+++ src/protocol/clientbound_keepalive.h
@@ -30,5 +30,5 @@ class ClientboundKeepAlive : public ClientboundPacket {
   [[nodiscard]] int getPacketID() const override;
 
  public:
-  long long id = 0;
+  int64_t id = 0;
 };
