+++ src/time.cpp
@@ -20,7 +20,7 @@
 
 #include <chrono>
 
-long long currentTimeMillis() {
+int64_t currentTimeMillis() {
   return std::chrono::duration_cast<std::chrono::milliseconds>(
              std::chrono::system_clock::now().time_since_epoch()).count();
 }
